feat: delete answer and question use case and test

+++ src/domain/forum/application/use-cases/delete-answer-comment.ts
@@ -2,6 +2,7 @@ import { AnswerCommentsRepository } from "../repositories/answer-comments-reposi
 
 interface DeleteAnswerCommentUseCaseRequest {
   answerCommentId: string;
+  authorId: string;
 }
 
 interface DeleteAnswerCommentUseCaseResponse {}
@@ -11,12 +12,17 @@ export class DeleteAnswerCommentUseCase {
 
   async execute({
     answerCommentId,
+    authorId,
   }: DeleteAnswerCommentUseCaseRequest): Promise<DeleteAnswerCommentUseCaseResponse> {
     const answerComments =
       await this.answerCommentsRepository.findById(answerCommentId);
 
     if (!answerComments) {
       throw new Error("Answer Comment not found.");
+    }
+
+    if (answerComments.authorId.toString() !== authorId) {
+      throw new Error("Not allowed");
     }
 
     await this.answerCommentsRepository.delete(answerComments);
